Remove const data members
Const data members prevent copy or move assignment operators from being
generated by the compiler. This is because the compiler-generated
assignment operators are not allowed to modify const data members.
It is considered poor practice to use const data members in C++,
unless for the express purpose of preventing copy or move assignment.